requirements.txt
tifffile>=2024.5.22
zarr>=3.0
fastapi>=0.104
uvicorn[standard]>=0.24
pandas>=2.0
pyarrow>=13.0
numpy>=1.24
scipy>=1.11
Pillow>=10.0
openpyxl>=3.1
pytest>=8.0